CREDITS.md
Upon registration a paired member receives a Multiplicity token which is a unique identifier that is used to unlock their membership. 

Sustaining Multiplicity does not so much depend on the certainty of its conclusion as its self corrective character.

Multiplicity is only recognized by this organization and by its members and is unlike any other form of currency, monetary instrument, or credit system. Multiplicity serves as the primary means by which we recognize and reward our Members for their participation. Our membership services can only be acquired by using Multiplicity in accordance with these rules and guidelines which govern our memberships and are outlined in the articles of incorporation and these bylaws. Every member is rewarded for their participation as follows:

In Summary

INGENUITY AWARDS, are the most esteemed and sought after awards available to any and all members. As a Research & Development Organization we are constantly looking for ways to optimize our resources so that we can multiply our productivity and our services to the greater community. One of the ways we do this is by rewarding our members for their ingenuity and creativity.

DESIGNATION REWARDS are incurred monthly by members who are rightfully able and willing to designate assets such as land, facilities, or intellectual property to be under the stewardship of the organization and used only for purposes approved by the designating member and the Board of Directors. Designation credits are rewarded according to the services and productivity that is a direct result of the designation and are typically rewarded on a monthly basis however in certain instances designation rewards may be rewarded on a weekly basis upon resolution of the board.

OPPORTUNITY REWARDS nothing works if we don’t, so our aim is to continuously have a wide variety of opportunities blended between time slots and full time positions. It takes people to get things done and it‘s to the honor of our hard working volunteers,  associates,  managers, and elected officials that we recognize and reward them for their time and effort. Opportunity credits can also be produced and exchanged between members. Opportunity rewards are distributed on weekly basis and are incurred based on the value of reciprocity provided by the opportunity.

SHARING REWARDS  are annually rewarded credits that any and all members can receive by inviting 2 new members. For the first referral a member will receive an annual 1% credit based on the total credits incurred by the first referred account and a .5% credit for the second referred account, for as long as the memberships stay active.  Sharing Rewards are distributed within 90 days from the first of the year.

SPONSORSHIP CREDITS always represent the value of a monetary gift, and do not incur interest however they are granted according to the law of Multiplicity. The purpose of community is to create a greater value within in order to the share the benefits with all.